refactor: read catalog metadata through the CWM helpers
Element texts go through Descriptions with the CwmHelper vocabulary,
catalog exposure through Packages.available, and the bespoke
rolap.mapping Annotation is replaced by CWM TaggedValues. RolapMetaData
is built from those and covered by a test.

The rest tracks the changed rolap.mapping API: SqlStatement carries a
body instead of a sql string, ExplicitAggregationTable carries its table
reference, and MemberLike is the common measure/member type. The core
run bundles gain cwm.foundation.businessinformation.

cwm.supplier exposes the rolap catalog as a CWM schema. It stays out of
the reactor until org.eclipse.daanse.olap.cwm.spi is published.

+++ core/src/main/java/org/eclipse/daanse/rolap/aggregator/extra/NthValueAggregator.java
@@ -28,11 +28,11 @@ import org.eclipse.daanse.sql.statement.api.expression.SqlExpression;
 public class NthValueAggregator implements Aggregator, org.eclipse.daanse.rolap.aggregator.SqlNodeAggregator {
 
     private boolean ignoreNulls;
-    private Integer n;
+    private long n;
     private List<RolapColumn> rolapOrderedColumnList;
 
 
-    public NthValueAggregator(boolean ignoreNulls, Integer n,
+    public NthValueAggregator(boolean ignoreNulls, long n,
                               List<RolapColumn> rolapOrderedColumnList) {
         this.ignoreNulls = ignoreNulls;
         this.n = n;
@@ -50,7 +50,7 @@ public class NthValueAggregator implements Aggregator, org.eclipse.daanse.rolap.
                 .toList();
         }
         return new SqlExpression.ExtraAggregate(Optional.ofNullable(operand),
-                new SqlExpression.ExtraAggregate.Spec.NthValue(ignoreNulls, n, columnsList));
+                new SqlExpression.ExtraAggregate.Spec.NthValue(ignoreNulls, (int) n, columnsList));
     }
 
     @Override
